WinRT.Module.GetActivationFactory for authoring/hosting integration
