+++ HostTest/TestHost.idl
@@ -1,15 +1,20 @@
 // Modern IDL 3.0: https://docs.microsoft.com/en-us/uwp/midl-3/intro
 
-// Definitions for testing C# runtime class hosting, not implemented in TestComponentCSharp
+// Definitions for testing C# runtime class hosting
 namespace TestHost
 {
-    runtimeclass ManagedClass : [default] Windows.Foundation.IStringable
+    runtimeclass ProbeByClass : [default] Windows.Foundation.IStringable
     {
-        ManagedClass();
+        ProbeByClass();
     }
 
-    runtimeclass ManagedClassNotFound : [default] Windows.Foundation.IStringable
+    runtimeclass ProbeByHost : [default] Windows.Foundation.IStringable
     {
-        ManagedClassNotFound();
+        ProbeByHost();
+    }
+
+    runtimeclass ClassNotFound : [default] Windows.Foundation.IStringable
+    {
+        ClassNotFound();
     }
 }
