--- Projections/Test/TestHost.ManagedClass.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Reflection;
-using System.Runtime.InteropServices;
-using System.Text;
-using Windows.Foundation;
-using WinRT;
-
-
-namespace TestHost
-{
-    public class ManagedClass : IStringable
-    {
-        public override string ToString() => "ManagedClass";
-    }
-}
