+++ Projections/TestHost.ProbeByClass/TestHost.ProbeByClass.csproj
@@ -1,0 +1,9 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <AssemblyName>TestHost.ProbeByClass</AssemblyName>
+  </PropertyGroup>
+
+  <Import Project="../Test/Test.csproj"/>
+
+</Project>
